+++ features/include/pcl/features/impl/pfhrgb.hpp
@@ -69,7 +69,7 @@ pcl::PFHRGBEstimation<PointInT, PointNT, PointOutT>::computePointPFHRGBSignature
   pfhrgb_histogram.setZero ();
 
   // Factorization constant
-  float hist_incr = 100.0 / (indices.size () * indices.size () - 1);
+  float hist_incr = 100.0f / static_cast<float> (indices.size () * indices.size () - 1);
 
   // Iterate over all the points in the neighborhood
   for (size_t i_idx = 0; i_idx < indices.size (); ++i_idx)
@@ -87,28 +87,28 @@ pcl::PFHRGBEstimation<PointInT, PointNT, PointOutT>::computePointPFHRGBSignature
         continue;
 
       // Normalize the f1, f2, f3, f5, f6, f7 features and push them in the histogram
-      f_index_[0] = floor (nr_split * ((pfhrgb_tuple_[0] + M_PI) * d_pi_));
+      f_index_[0] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[0] + M_PI) * d_pi_)));
       if (f_index_[0] < 0)         f_index_[0] = 0;
       if (f_index_[0] >= nr_split) f_index_[0] = nr_split - 1;
 
-      f_index_[1] = floor (nr_split * ((pfhrgb_tuple_[1] + 1.0) * 0.5));
+      f_index_[1] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[1] + 1.0) * 0.5)));
       if (f_index_[1] < 0)         f_index_[1] = 0;
       if (f_index_[1] >= nr_split) f_index_[1] = nr_split - 1;
 
-      f_index_[2] = floor (nr_split * ((pfhrgb_tuple_[2] + 1.0) * 0.5));
+      f_index_[2] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[2] + 1.0) * 0.5)));
       if (f_index_[2] < 0)         f_index_[2] = 0;
       if (f_index_[2] >= nr_split) f_index_[2] = nr_split - 1;
 
       // color ratios are in [-1, 1]
-      f_index_[4] = floor (nr_split * ((pfhrgb_tuple_[4] + 1.0) * 0.5));
+      f_index_[4] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[4] + 1.0) * 0.5)));
       if (f_index_[4] < 0)         f_index_[4] = 0;
       if (f_index_[4] >= nr_split) f_index_[4] = nr_split - 1;
 
-      f_index_[5] = floor (nr_split * ((pfhrgb_tuple_[5] + 1.0) * 0.5));
+      f_index_[5] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[5] + 1.0) * 0.5)));
       if (f_index_[5] < 0)         f_index_[5] = 0;
       if (f_index_[5] >= nr_split) f_index_[5] = nr_split - 1;
 
-      f_index_[6] = floor (nr_split * ((pfhrgb_tuple_[6] + 1.0) * 0.5));
+      f_index_[6] = static_cast<int> (floor (nr_split * ((pfhrgb_tuple_[6] + 1.0) * 0.5)));
       if (f_index_[6] < 0)         f_index_[6] = 0;
       if (f_index_[6] >= nr_split) f_index_[6] = nr_split - 1;
